requirements/base.txt
django==4.1
django-environ==0.9.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
drf-yasg

Pillow==9.0.0

argon2-cffi==21.3.0

pytz==2021.3

redis==4.1.0
celery==5.2.3
flower==1.0.0

PyJWT
djangorestframework-simplejwt
